Version 1
This is the first version of the project.

There is some errors in the data collection in getting the SQFT of the property listings, but this was corrected internally by deriving the SQFT from the Price and Price per SQFT, which were gathered correctly.

While this version is already workable, there are a lot of potential improvements when it comes to the accuracy of the machine learning model (R) and design of the dashboard (Tableau). In addition, bug fixes could be made in Python code, which would reduce the effort required for data cleaning in latter stages.